Swap names of ClientProtocol and Peer
